Add _layouts/default.html and use it.
This fixes Jon's issue with clickability and it is worth moving most of the
world into _layouts as html fragments. The current approach of duplicating
keeps things very straightforward, which is good for having this be a project
that anyone can contribute to, but it's getting to be too much.